Move odds primitives to odds-core, drop odds-lambda→odds-analytics dep
Dockerfile.scraper has been failing since #327 because odds-lambda declared
odds-analytics as a hard dependency, but the scraper image intentionally does
not bundle odds-analytics (too heavy). `uv pip install odds_lambda-*.whl` then
failed with "odds-analytics was not found in the package registry".

Move the lambda-runtime primitives that lived in odds-analytics down into
odds-core so odds-lambda can import them without pulling in the analytics
package:

- odds_core.odds_math: american/decimal conversion, implied probability,
  devigging, h2h winner, profit-from-odds.
- odds_core.snapshot_utils.extract_odds_from_snapshot: parses
  OddsSnapshot.raw_data into Odds rows.

odds-analytics keeps its analytics-specific helpers (risk metrics, arbitrage
detection, tier coverage table) and re-imports the primitives from odds-core.
This also breaks the odds-lambda ⇄ odds-analytics cycle.

Also move packages/odds-mcp/tests/test_server.py → tests/unit/test_mcp_server.py
so CI (`pytest tests/unit`) actually runs it — it has been invisible since
#273. Fix the two tests that silently went stale in the meantime:

- test_known_league_uses_correct_sport: refresh_scrape stopped calling
  ingest_league directly in #328 and now submits via add_job; assert against
  the scheduler mock instead.
- test_missing_event (TestGetCurrentOddsMissingEvent): get_current_odds
  started requiring `market` in #304.
