+++ packages/odds-analytics/odds_analytics/__init__.py
@@ -8,15 +8,16 @@ install with ``pip install odds-analytics[training]`` for full functionality.
 Modules that require training extras raise ImportError at import time if missing.
 """
 
-from odds_analytics.sequence_loader import load_sequences_for_event
-from odds_analytics.utils import (
+from odds_core.odds_math import (
     american_to_decimal,
-    calculate_ev,
     calculate_implied_probability,
     calculate_profit_from_odds,
     decimal_to_american,
     determine_h2h_winner,
 )
+
+from odds_analytics.sequence_loader import load_sequences_for_event
+from odds_analytics.utils import calculate_ev
 
 
 def __getattr__(name: str) -> object:
