+++ packages/odds-lambda/pyproject.toml
@@ -6,7 +6,6 @@ authors = [{name = "Andrew"}]
 requires-python = ">=3.12"
 dependencies = [
     "odds-core",
-    "odds-analytics",
     "aiohttp>=3.9.0",
     "apscheduler==4.0.0a6",
     "tenacity>=8.2.3",
@@ -21,7 +20,6 @@ dependencies = [
 
 [tool.uv.sources]
 odds-core = { workspace = true }
-odds-analytics = { workspace = true }
 
 [build-system]
 requires = ["hatchling"]
